longs in mainwindow have been changed to qint32

+++ trunk/client/mainwindow.h
@@ -36,7 +36,7 @@ private slots:
     void on_actionBlackPen_triggered();
     void on_actionRedPen_triggered();
     void on_actionEraser_triggered();
-    void sessionJoinResult(QString sessionName, QChar result, QHash<QString, long> users);
+    void sessionJoinResult(QString sessionName, QChar result, QHash<QString, qint32> users);
     void drawingCommitted(QString sessionName, QPicture pictureData);
     void on_actionStraightLine_triggered();
     void on_actionRectangle_triggered();
